# This is an experimental branch that might lead to some faster performance when you expect long significand (more than 19 digits)