--- a/include/fast_float/parse_number.h
+++ b/include/fast_float/parse_number.h
@@ -91,17 +91,48 @@ from_chars_result from_chars(const char *first, const char *last,
   answer.ec = std::errc(); // be optimistic
   answer.ptr = pns.lastmatch;
 
-  if (binary_format<T>::min_exponent_fast_path() <= pns.exponent && pns.exponent <= binary_format<T>::max_exponent_fast_path() && pns.mantissa <=binary_format<T>::max_mantissa_fast_path()) {
-    value = T(pns.mantissa);
-    if (pns.exponent < 0) { value = value / binary_format<T>::exact_power_of_ten(-pns.exponent); } 
-    else { value = value * binary_format<T>::exact_power_of_ten(pns.exponent); }
-    if (pns.negative) { value = -value; }
-    return answer;
+  answer.ec = std::errc(); // be optimistic
+  answer.ptr = pns.lastmatch;
+  adjusted_mantissa am;
+  // Most times, we have pns.too_many_digits = false.
+  if(pns.too_many_digits) {
+    // Uncommon path where we have too many digits.
+    //
+    // credit: R. Oudompheng who first implemented this fast path.
+    // It does the job of accelerating the slow path since most
+    // long streams of digits are determined after 19 digits.
+    // Note that mantissa+1 cannot overflow since mantissa < 10**19 and so
+    // mantissa+1 <= 10**19 < 2**64.
+    adjusted_mantissa am1 = compute_float<binary_format<T>>(pns.exponent, pns.mantissa);
+    adjusted_mantissa am2 = compute_float<binary_format<T>>(pns.exponent, pns.mantissa+1);
+    // They must both agree and be both a successful result.
+    if(( am1 == am2 ) && (am1.power2 >= 0)) {
+      am = am1;
+    } else {
+      // long way! (uncommon)
+      decimal d = parse_decimal(first, last);
+      am = compute_float<binary_format<T>>(d);
+    }
+  } else {
+    // We are entering the common path where the number of digits is no more than 19.
+    //
+    // Next is Clinger's fast path.
+    if (binary_format<T>::min_exponent_fast_path() <= pns.exponent && pns.exponent <= binary_format<T>::max_exponent_fast_path() && pns.mantissa <=binary_format<T>::max_mantissa_fast_path()) {
+      value = T(pns.mantissa);
+      if (pns.exponent < 0) { value = value / binary_format<T>::exact_power_of_ten(-pns.exponent); }
+      else { value = value * binary_format<T>::exact_power_of_ten(pns.exponent); }
+      if (pns.negative) { value = -value; }
+      return answer;
+    }
+    // Then we have our main routine.
+    am = compute_float<binary_format<T>>(pns.exponent, pns.mantissa);
+    // If we called compute_float<binary_format<T>>(pns.exponent, pns.mantissa) and we have an invalid power (am.power2 < 0),
+    // then we need to go the long way around again. This is very uncommon.
+    if(am.power2 < 0) { // long way! (uncommon)
+      decimal d = parse_decimal(first, last);
+      am = compute_float<binary_format<T>>(d);
+    }
   }
-  adjusted_mantissa am = pns.too_many_digits ? parse_long_mantissa<binary_format<T>>(first,last) : compute_float<binary_format<T>>(pns.exponent, pns.mantissa);
-  // If we called compute_float<binary_format<T>>(pns.exponent, pns.mantissa) and we have an invalid power (am.power2 < 0),
-  // then we need to go the long way around again. This is very uncommon.
-  if(am.power2 < 0) { am = parse_long_mantissa<binary_format<T>>(first,last); }
   uint64_t word = am.mantissa;
   word |= uint64_t(am.power2) << binary_format<T>::mantissa_explicit_bits();
   word = pns.negative